+++ MaskGuestBook/admin.py
@@ -2,7 +2,4 @@ from django.contrib import admin
 from .models import *
 
 # Register your models here.
-admin.site.register(GuestBook)
-admin.site.register(Image)
 admin.site.register(GuestBookModel)
-
