+++ MaskGuestBook/static/css/GuestBook.css
@@ -80,6 +80,3 @@ h4 {
     margin-right: 15px;
     margin-top: 15px;
 }
-
-/*부모 : relative*/
-/*자식 : absolute, top, bottom...*/
